Move Converter and BidirectionalConverter of JavaFX client API to com.canoo.platform.remoting.client.javafx Fix typo in Converter JavaDoc Remove BidirectionalConverter import from Converter
Fixes #635
