+++ platform/dolphin-platform-remoting-client-javafx/src/main/java/com/canoo/dp/impl/client/javafx/AbstractNumericDolphinBinder.java
@@ -17,7 +17,7 @@ package com.canoo.dp.impl.client.javafx;
 
 import com.canoo.platform.core.functional.Subscription;
 import com.canoo.platform.remoting.Binding;
-import com.canoo.platform.remoting.client.javafx.binding.BidirectionalConverter;
+import com.canoo.platform.remoting.client.javafx.BidirectionalConverter;
 import com.canoo.platform.remoting.client.javafx.binding.NumericDolphinBinder;
 import com.canoo.platform.remoting.Property;
 import javafx.beans.value.ChangeListener;
