+++ platform/dolphin-platform-remoting-client-javafx/src/main/java/com/canoo/dp/impl/client/javafx/DefaultDolphinBinder.java
@@ -15,9 +15,9 @@
  */
 package com.canoo.dp.impl.client.javafx;
 
-import com.canoo.platform.remoting.client.javafx.binding.BidirectionalConverter;
+import com.canoo.platform.remoting.client.javafx.BidirectionalConverter;
 import com.canoo.platform.remoting.Binding;
-import com.canoo.platform.remoting.client.javafx.binding.Converter;
+import com.canoo.platform.remoting.client.javafx.Converter;
 import com.canoo.platform.remoting.client.javafx.binding.DolphinBinder;
 import com.canoo.platform.core.functional.Subscription;
 import com.canoo.platform.remoting.Property;
