+++ platform/dolphin-platform-remoting-client-javafx/src/main/java/com/canoo/dp/impl/client/javafx/DefaultJavaFXBinder.java
@@ -16,7 +16,7 @@
 package com.canoo.dp.impl.client.javafx;
 
 import com.canoo.platform.remoting.Binding;
-import com.canoo.platform.remoting.client.javafx.binding.Converter;
+import com.canoo.platform.remoting.client.javafx.Converter;
 import com.canoo.platform.remoting.client.javafx.binding.JavaFXBinder;
 import com.canoo.platform.core.functional.Subscription;
 import com.canoo.platform.remoting.Property;
