+++ platform/dolphin-platform-remoting-client-javafx/src/test/java/com/canoo/dolphin/client/javafx/FXBinderTest.java
@@ -15,9 +15,9 @@
  */
 package com.canoo.dolphin.client.javafx;
 
-import com.canoo.platform.remoting.client.javafx.binding.BidirectionalConverter;
+import com.canoo.platform.remoting.client.javafx.BidirectionalConverter;
 import com.canoo.platform.remoting.Binding;
-import com.canoo.platform.remoting.client.javafx.binding.Converter;
+import com.canoo.platform.remoting.client.javafx.Converter;
 import com.canoo.platform.remoting.client.javafx.binding.DefaultBidirectionalConverter;
 import com.canoo.platform.remoting.client.javafx.FXBinder;
 import com.canoo.platform.remoting.ObservableList;
